Removed new conftest fixtures and refactored to use the imputer test_data fixture.  Moved the imputer_test_data fixture up into conftest.py so that test_simple_imputer could borrow it.
